same as last sdadc ok <1mv

+++ MFC/MFC-MCUf373/hardware/adc/mfc_adc.h
@@ -6,10 +6,10 @@
  
 uint32_t SDADC1_Config(void);
 void GetPosition(void);
-unsigned int GetADCVoltage(unsigned char ch);
+float GetADCVoltage(unsigned char ch);
 void SetVotageTimes(unsigned int val);
-int16_t ADC_Filter(unsigned char ch) ;
-int16_t ADC_Mean(unsigned char ch);
+float ADC_Filter(unsigned char ch) ;
+float ADC_Mean(unsigned char ch);
 
 #endif /* __ADC_H */
